Add some safety for cases when killed before really being alive.

+++ src/pythonKarabo/karabo/bound_api/device_server.py
@@ -313,7 +313,8 @@ class DeviceServer(object):
 
     def stopDeviceServer(self):
         self.scanning = False
-        self.pluginThread.join()
+        if hasattr(self, "pluginThread") and self.pluginThread.isAlive():
+            self.pluginThread.join()
         self.pluginThread = None
         self.ss = None
         EventLoop.stop()
@@ -411,7 +412,10 @@ class DeviceServer(object):
         self.log.WARN("DeviceServer \"{}\" does not allow the transition for this event.".format(self.serverid))
 
     def slotKillServer(self):
-        self.log.INFO("Received kill signal")
+        if self.log:
+            self.log.INFO("Received kill signal")
+        else: # might get killed by signal handler before setting up logging
+            print("Received kill signal")
         launchers = []
         for deviceid in list(self.deviceInstanceMap.keys()):
             self.ss.call(deviceid, "slotKillDevice")
@@ -423,9 +427,10 @@ class DeviceServer(object):
         try:
             self.ss.reply(self.serverid)
         except Exception as e:
-            msg = ("Did not notify distributed system of server shutdown:"
-                   "\n {}").format(e)
-            self.log.ERROR(msg)
+            if self.log: # see above
+                msg = ("Did not notify distributed system of server shutdown:"
+                       "\n {}").format(e)
+                self.log.ERROR(msg)
         finally:
             # NOTE: `stopDeviceServer` will not return because the process will
             # be killed - everything which is called after this call will
